Add metadata support per route
