+++ webapp/app/layout.tsx
@@ -1,8 +1,31 @@
 import { locales } from 'app/i18n'
+import { Metadata } from 'next'
 import { ReactNode } from 'react'
 
 type Props = {
   children: ReactNode
+}
+
+export const metadata: Metadata = {
+  description:
+    'Hemi | Powered by Bitcoin and Ethereum — Bridge, swap, build, and more. All on Hemi.',
+  openGraph: {
+    description:
+      'Hemi | Powered by Bitcoin and Ethereum — Bridge, swap, build, and more. All on Hemi.',
+    images: [
+      {
+        url: 'https://hemi.xyz/image/portal.png',
+      },
+    ],
+    type: 'website',
+  },
+  title: 'Hemi Portal',
+  twitter: {
+    card: 'summary_large_image',
+    description:
+      'Hemi | Powered by Bitcoin and Ethereum — Bridge, swap, build, and more. All on Hemi.',
+    images: ['https://hemi.xyz/image/portal.png'],
+  },
 }
 
 export const generateStaticParams = async () =>
